requirements.txt
streamlit
requests
sentence-transformers
faiss-cpu
openai
python-dotenv
